get2gnow #0000ECaFC - ...abunch more...for later...
After fixing my last bug, my first bug started randomly accuring again.  I've finally resolved the issue by switching from float to double values for update & user ids.  I've had to store the values in gconf as a string & use strtod when I pull the values out again.  I know its a hack but its the only way to avoid issues that come up when using floats.  For what ever reason float values weren't converted from stings correctly.  It was literally coming up with a completely different value.

I've alo implement libgnome & using GnomProgram.  I've done this for many  future proofing reasons but right now my primary new feaure why I've implemented GnomeProgam right now is to have get2gnow run in 'C' locale, before threading is started, and than to ensure the orginal locale is always reset once get2gnow exits.  This is so multiple date, time, & parting methods are all now thread safe.

Lastly I'm increased the persion of get2gnow sorting and other time / date values used by get2gnow for sorting and time stamping.  Basically I've added a check so now new updates that end up calulating to less than 1 second ago, which can happen, and now when it does get2gnow's sorting value will be set to a minimum of 1 to ensure even the abs newest updates appear in the correct order.

Lastly, I've update 'SelectedTweet' actions, line star'ing, unstar'ing, following, & etc to use gdouble values as well.  Now all of those actions are working again.

A note reguarding using the new gdouble values as strings.  I've created a method to drop precision from a gdouble value and than returns its value, w/o zeros, as a gchar *.  My function is called 'program_drop_precision' but its intended to be called using my macro 'gdouble_to_str'.

There, no all shiny, all done, & now I disconnect from this world of 0s & 1s to go back to my life of one...oh my oh my what fun.

Oh yeah uber coolness note, its gonna make me hella tight come next month & prolly into Sept, but it was def w/o any doubt 100

Now I'm just giddily waiting & giddily working again.  I'll soon no longer feel afraid to leave my bed, I can't.  As for why I've felt afraid to leave my bed well afraid isn't the right word, its that how alone I am really hits home then.  The worst part is that we're left w/no distrattions, noise, or company, only our thoughts & memories.  Which is an istruction set for periods of pure hell & that's not an exaggeration in fact its the only way we could possible describe what its like...its what most of our life is like.  But it is getting better...

Its bizarre, life was hell for so long just pure hell.  Than I fought for 6-8yrs to get to a better place w/a better life.  & I did it, I'd won, I was living, happy, hopeful, & working toward my future w/all of my dreams seemingly coming true.  Than out of no where Generalized Dystonia started & over a few short years despite fighting to hold on to what I could I eventually lost every thing.  And I do truly mean every thing, we lost it all, every possetion, our jobs, our abbility to drive, our apt, our friends, our cats, every thing but us.  After loosing every thing I then found myself even needing a care taker, and yes I truly need my PCAs.  Its weird, only as I'm writing am I realizing something.  In November of 2007 I went in for the last procedure that, for us, had offerred any hope.  Well whe it didn't help - it didn't work.  And since than I've stopped trying to get any thing back, I've stopped trying to build my life again, I'd stopped hoping.  Why?  Because I'm afraid of loosing every thing all over again but still being left alive.  Loosing every thing expect sufferring is trying worse than death.  Its not a whole I want to ever fall back down, of course I'll have to crawl out 1st.  Well its only writing this that I've gained this insight on why I feel how I feel.  And now knowing it, understanding it, well I'm ready to start fighting like hell again.  I'll have to do it every day & forever, but I'm ready.  So we're now deciding to start re-building & to fight the good fight.  Now's our time to get beck to our lives, to get back to my life, to start getting thing back, putting them together, & to start living...once again.  Our new & current primary project: refactoring my life.  1st step getting read for bed then sleeping.

Committed on: 2009-06-27 @ 12:39:14 AM

+++ src/online-service-request.h
@@ -101,11 +101,11 @@ void online_service_request_selected_tweet_unblock(void);
 void online_service_request_selected_tweet_save_fave(void);
 void online_service_request_selected_tweet_destroy_fave(void);
 
-void set_selected_tweet(OnlineService *service, const gfloat id, const gfloat user_id, const gchar *user_name, const gchar *tweet);
+void set_selected_tweet(OnlineService *service, const gdouble id, const gdouble user_id, const gchar *user_name, const gchar *tweet);
 OnlineService *selected_tweet_get_service(void);
-gfloat selected_tweet_get_id(void);
+gdouble selected_tweet_get_id(void);
 gchar *selected_tweet_get_user_name(void);
-gfloat selected_tweet_get_user_id(void);
+gdouble selected_tweet_get_user_id(void);
 gchar *selected_tweet_reply_to_strdup(gboolean retweet);
 void selected_tweet_reply(void);
 void selected_tweet_retweet(void);
